cleanup plotter
