+++ src/cosmic_integration/observation.py
@@ -2,6 +2,8 @@ from dataclasses import dataclass
 
 import numpy as np
 
+from .utils import read_output
+
 
 @dataclass
 class Observation:
@@ -12,17 +14,7 @@ class Observation:
 
     @classmethod
     def from_jeff(self, fname: str, idx: int = 0) -> 'Observation':
-        data = np.genfromtxt(fname, delimiter=',')
-
-        if data.ndim == 1:
-            # If the data is 1D, reshape it to a 2D array with one row
-            data = data.reshape(1, -1)
-
-        params = data[idx, :4]
-        shape = tuple(int(x) for x in data[idx, 4:6])
-        ndata = shape[0] * shape[1]
-
-        matrix = data[idx, 6:6 + ndata].reshape(shape)
+        matrix, params, _ = read_output(fname, idx)
         return Observation(
             duration=1,
             rate_matrix=matrix,
